requirements.txt
google-cloud-bigquery=3.13.0
requests=2.31.0
flask=3.0.0
matplotlib=3.7.3
pandas=2.1.2
geopandas=0.14.0
shapely=2.0.1
cartopy=0.22.0
python-dotenv=1.0.1
basemap=1.3.7
